format alumni & profile tables

+++ src/main/java/org/alica/api/service/AlumniService.java
@@ -62,8 +62,7 @@ public class AlumniService {
     }
 
     public void deleteAlumni(UUID id){
-        Alumni alumni = alumniRepository.findById(id).orElseThrow(() -> new PropertyNotFoundException(String.format("Alumni %s Not found !",id)));
+        if(!alumniRepository.existsById(id)) throw new PropertyNotFoundException(String.format("Alumni %s Not found !",id));
         alumniRepository.deleteById(id);
     }
-
 }
